+++ src/frontend/Cargo.toml
@@ -32,6 +32,7 @@ web-sys = { version = "0.3.77", features = ['Headers','CustomEvent','CustomEvent
 winit = { version = "0.30", features = ["rwh_05"] }
 bytemuck = { version =  "1.21.0", features = ["derive"] }
 glam = "0.30"
+geo ="*"
 web-time= "1.1"
 image="0.25"
